requirements.txt
# ── Core RAG stack ──────────────────────────────────────────
langchain>=0.2.0,<0.4.0
langchain-community>=0.2.0,<0.4.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
sentence-transformers>=2.2.2

# ── LLM provider ───────────────────────────────────────────
mistralai>=2.0.0

# ── Utilities ──────────────────────────────────────────────
python-dotenv>=1.0.0